Rework the launcher arguments parsing
