+++ services/start-import
@@ -1,4 +1,4 @@
 #!/bin/sh
 set -eu
-"$LEARN2RAG_PATH"/configurator learn2rag.pipeline.importer --config "$STORAGE_PATH"/importer_config.json
-"$LEARN2RAG_PATH"/configurator learn2rag.pipeline.ingestion
+"$LEARN2RAG_PATH"/configurator learn2rag.importer --config "$STORAGE_PATH"/importer_config.json
+"$LEARN2RAG_PATH"/configurator learn2rag.pipeline.ingestion --logging-config "$STORAGE_PATH"/logging_config.yml
