+++ services/start-import.windows.bat
@@ -1,2 +1,2 @@
-"%LEARN2RAG_PATH%/configurator.exe" learn2rag.pipeline.importer --config "%STORAGE_PATH%/importer_config.json"
+"%LEARN2RAG_PATH%/configurator.exe" learn2rag.importer --config "%STORAGE_PATH%/importer_config.json"
 "%LEARN2RAG_PATH%/configurator.exe" learn2rag.pipeline.ingestion
